requirements.txt
quart
quart-cors
httpx
requests
python-dotenv
coloredlogs
APScheduler
hypercorn

